// src/interface.d.ts
export interface IProduct{
    id:number
    name:string
    manufacturer:string
    type:string
    category:string
    price:number
    discount?:number
    aroundChest?:number
    height?:number
    totalPrice?:number 
    count?:any
    roundShip?:number
    soleFoot?:number
    noSize?:boolean
    waist?:number
    img:string
    subImg:string
}